separate useCalculations and useValidations
calculate only currently table dependant nodes for changed cell
