+++ src/client/store/ui/assessmentSection/slice.ts
@@ -43,7 +43,24 @@ export const assessmentSectionSlice = createSlice({
         state.data = TableDatas.updateDatum({ data: state.data, countryIso, tableName, variableName, colName, value })
       })
     },
+
     setNodeValidations: (state, action: PayloadAction<{ nodeUpdates: NodeUpdates }>) => {
+      const { nodeUpdates } = action.payload
+      const { countryIso, nodes } = nodeUpdates
+
+      nodes.forEach(({ tableName, variableName, colName, value }) => {
+        state.data = TableDatas.updateDatumValidation({
+          data: state.data,
+          countryIso,
+          tableName,
+          variableName,
+          colName,
+          validation: value.validation,
+        })
+      })
+    },
+
+    setNodeCalculations: (state, action: PayloadAction<{ nodeUpdates: NodeUpdates }>) => {
       const { nodeUpdates } = action.payload
       const { countryIso, nodes } = nodeUpdates
